+++ src/OpenTelemetry.Contrib.Exporter.ElasticApm/Implementation/V2/Metadata.cs
@@ -1,6 +1,4 @@
-﻿using System.Diagnostics;
-using System.Reflection;
-using System.Text.Json;
+﻿using System.Text.Json;
 
 namespace OpenTelemetry.Contrib.Exporter.Elastic.Implementation.V2
 {
@@ -24,57 +22,6 @@ namespace OpenTelemetry.Contrib.Exporter.Elastic.Implementation.V2
             this.Service.Write(writer);
 
             writer.WriteEndObject();
-
-            writer.WriteEndObject();
-        }
-    }
-
-    internal readonly struct Service
-    {
-        public Service(string name, string environment, Agent agent)
-        {
-            this.Name = name;
-            this.Environment = environment;
-            this.Agent = agent;
-        }
-
-        internal string Name { get; }
-
-        internal string Environment { get; }
-
-        internal Agent Agent { get; }
-
-        public void Write(Utf8JsonWriter writer)
-        {
-            writer.WriteStartObject();
-
-            writer.WriteString(JsonHelper.NamePropertyName, this.Name);
-            writer.WriteString(JsonHelper.EnvironmentPropertyName, this.Environment);
-            writer.WritePropertyName(JsonHelper.AgentPropertyName);
-            this.Agent.Write(writer);
-
-            writer.WriteEndObject();
-        }
-    }
-
-    internal readonly struct Agent
-    {
-        public Agent(Assembly assembly)
-        {
-            this.Name = "opentelemetry";
-            this.Version = FileVersionInfo.GetVersionInfo(assembly.Location).ProductVersion;
-        }
-
-        internal string Name { get; }
-
-        internal string Version { get; }
-
-        public void Write(Utf8JsonWriter writer)
-        {
-            writer.WriteStartObject();
-
-            writer.WriteString(JsonHelper.NamePropertyName, this.Name);
-            writer.WriteString(JsonHelper.VersionPropertyName, this.Version);
 
             writer.WriteEndObject();
         }
